refactor: move tests to components
