+++ src/components/list-result/Card.test.tsx
@@ -1,26 +1,12 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
-import Card from '../components/list-result/Card';
-import App from '../App';
-import { ThemeProvider } from '../context/ThemeContext';
+import Card from './Card';
+import App from '../../App';
+import { ThemeProvider } from '../../context/ThemeContext';
 import { Provider } from 'react-redux';
-import { mockStore } from './mock/mockStore';
+import { mockStore } from '../../__tests__/mock/mockStore';
 import { MemoryRouter } from 'react-router-dom';
-
-const product = {
-  id: 1,
-  title: 'Iphone',
-  description: 'Description: SIM-Free,',
-  price: 1000,
-  discountPercentage: 10,
-  rating: 10,
-  stock: 4,
-  brand: 'apple',
-  category: 'phones',
-  thumbnail: ['moc, apple'],
-  species: ['red', 'blue'],
-  images: ['image.png'],
-};
+import { productMock } from '../../__tests__/mock/mockStore';
 
 describe('Card Component', () => {
   it('image is rendered', () => {
@@ -28,7 +14,7 @@ describe('Card Component', () => {
       <MemoryRouter>
         <ThemeProvider>
           <Provider store={mockStore}>
-            <Card product={product} page={1} />
+            <Card product={productMock} page={1} />
           </Provider>
         </ThemeProvider>
       </MemoryRouter>
@@ -40,7 +26,7 @@ describe('Card Component', () => {
       <MemoryRouter>
         <ThemeProvider>
           <Provider store={mockStore}>
-            <Card product={product} page={1} />
+            <Card product={productMock} page={1} />
           </Provider>
         </ThemeProvider>
       </MemoryRouter>
@@ -49,7 +35,7 @@ describe('Card Component', () => {
     expect(name).toBeInTheDocument();
   });
   it('Validate that clicking on a card opens a detailed card component', async () => {
-    localStorage.setItem('valueKey', product.title);
+    localStorage.setItem('valueKey', productMock.title);
     render(<App />);
 
     const name = await screen.findByText('Name: Essence Mascara Lash Princess');
