+++ src/store/reducers/selectedItems.slice.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { IProduct } from '../../types/types';
 import { selectedItemsSlice } from './selectedItems.slice';
-import { product } from '../../__tests__/Card.test';
+import { productMock } from '../../__tests__/mock/mockStore';
 
 describe('selectedItemsSlice', () => {
   const initialState = {
@@ -16,10 +16,10 @@ describe('selectedItemsSlice', () => {
   it('should handle selectItem', () => {
     const nextState = selectedItemsSlice.reducer(
       initialState,
-      selectedItemsSlice.actions.selectItem(product)
+      selectedItemsSlice.actions.selectItem(productMock)
     );
     expect(nextState.selectedItems).toHaveLength(1);
-    expect(nextState.selectedItems[0]).toEqual(product);
+    expect(nextState.selectedItems[0]).toEqual(productMock);
   });
 
   it('should handle unselectItem', () => {
@@ -28,18 +28,18 @@ describe('selectedItemsSlice', () => {
     };
     selectedItemsSlice.reducer(
       initialStateWithItems,
-      selectedItemsSlice.actions.selectItem(product)
+      selectedItemsSlice.actions.selectItem(productMock)
     );
     const nextState = selectedItemsSlice.reducer(
       initialStateWithItems,
-      selectedItemsSlice.actions.unselectItem(product)
+      selectedItemsSlice.actions.unselectItem(productMock)
     );
     expect(nextState.selectedItems).toHaveLength(0);
   });
 
   it('should handle clearSelectedItems', () => {
     const initialStateWithItems = {
-      selectedItems: [product] as IProduct[],
+      selectedItems: [productMock] as IProduct[],
     };
     const nextState = selectedItemsSlice.reducer(
       initialStateWithItems,
